requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.3
djangorestframework==3.14.0
drf-extra-fields==3.4.1
drf-writable-nested==0.7.0
mysqlclient==2.1.1
Pillow==9.3.0
pycodestyle==2.9.1
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6

